Fix for recalculate component height after resize
Now when you try to change the size of the component the height of the component is set as inline style,
which makes the recheck of the height not accurate because the height is in the style. you need to remove it first (set to 'auto') and recalc the element height.